+++ assets/js/blocks/cart-checkout/cart/empty-cart-edit/icon-data-uri.js
